Changed Feed.check to check any due before the next poll
Previously feeds would only be checked if a check was overdue.
However, this would mean that checks would always fall out of their
correct polling slot. For example, if they should be checked once
an hour with the first check at 10am, if the check run took 10 minutes
then some would get scheduled for a check at 11.10am - but they would
not be picked up at the 11am check, and would have to wait until 12am.

Now something scheduled for 11.10am will be checked at 11am, and
polling slots will not be missed.