Fixed race condition when changing feeds while scrolled
To recreate bug, enter a feed and scroll down to the bottom so all pks are
loaded, then switch to another feed. Sometimes the scroll event would trigger
a second call to Entries.loadNext(), but because pkUnloaded had already been
sent for loading and emptied, the message "No unread items" would be shown
before the original ajax call returned the items. This would shift the click
index, causing the wrong item to open, and raising an error on the last item.